models/requirements.txt
numpy
pandas
scikit-learn
matplotlib
flask
flask-cors
tensorflow
Pillow
